Валидация категорий админки

+++ app/Http/Controllers/Admin/ProductController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Category;
+use App\Http\Requests\ProductRequest;
 use App\Http\Controllers\Controller;
 use App\Product;
 use Illuminate\Http\Request;
@@ -38,7 +39,7 @@ class ProductController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(ProductRequest $request)
     {
         $path = $request -> file('image') -> store('products');
         $params = $request -> all();
@@ -77,12 +78,16 @@ class ProductController extends Controller
      * @param  \App\Product  $product
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Product $product)
+    public function update(ProductRequest $request, Product $product)
     {
-        Storage::delete($product -> image);
-        $path = $request -> file('image') -> store('products');
         $params = $request -> all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request ->has('image')) {
+            Storage::delete($product -> image);
+            $path = $request -> file('image') -> store('products');
+            $params['image'] = $path;
+        }
+
         $product -> update($params);
         return redirect()->route('products.index');
     }
